Split generate and test to separate mods.
