+++ tools/spectool/src/test.rs
@@ -1,0 +1,3 @@
+pub fn main() {
+    todo!()
+}
